fixed migration

+++ db/migrate/20130517162851_drop_food_tags.rb
@@ -1,6 +1,10 @@
 class DropFoodTags < ActiveRecord::Migration
   def up
-    drop_table :food_tags
-    # FoodTag.drop_translation_table!
+    begin
+      drop_table :food_tags if self.table_exists?("combo_types")
+      FoodTag.drop_translation_table!
+    rescue
+      
+    end
   end
 end
